+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Models/ResourceSchemas/ResourceOperationKind.cs
@@ -8,6 +8,8 @@ public enum ResourceOperationKind
 {
     /// <summary> A custom action operation. </summary>
     Action,
+    /// <summary> A check existence operation. </summary>
+    CheckExistence,
     /// <summary> A create or update operation. </summary>
     Create,
     /// <summary> A delete operation. </summary>