+++ sdk/ai/Azure.AI.Extensions.OpenAI/tests/Samples/Sample_HostedAgent.cs
@@ -88,12 +88,7 @@ public class Sample_HostedAgent : ProjectsOpenAITestBase
         Console.WriteLine($"The Agent {patchedRecord.Name} was patched.");
         #endregion
         #region Snippet:Sample_GetResponseFromAgentEndpoint_HostedAgent_Async
-        ProjectOpenAIClientOptions responsesOptions = new()
-        {
-            AgentName = agentVersion.Name
-        };
-        ProjectOpenAIClient openAIClient = new(uriEndpoint, credential, responsesOptions);
-        ProjectResponsesClient responseClient = openAIClient.GetProjectResponsesClient();
+        ProjectResponsesClient responseClient = projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(agentVersion.Name);
         ResponseResult response = await responseClient.CreateResponseAsync("Hello, tell me a joke.");
         Console.WriteLine(response.GetOutputText());
         #endregion
@@ -157,12 +152,7 @@ public class Sample_HostedAgent : ProjectsOpenAITestBase
         Console.WriteLine($"The Agent {patchedRecord.Name} was patched.");
         #endregion
         #region Snippet:Sample_GetResponseFromAgentEndpoint_HostedAgent_Sync
-        ProjectOpenAIClientOptions responsesOptions = new()
-        {
-            AgentName = agentVersion.Name
-        };
-        ProjectOpenAIClient openAIClient = new(uriEndpoint, credential, responsesOptions);
-        ProjectResponsesClient responseClient = openAIClient.GetProjectResponsesClient();
+        ProjectResponsesClient responseClient = projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(agentVersion.Name);
         ResponseResult response = responseClient.CreateResponse("Hello, tell me a joke.");
         Console.WriteLine(response.GetOutputText());
         #endregion